evio: stop accepting connections when close to descriptors rlimit
On connection, an evio service callback is invoked to accept it. In
some cases (like tarantool being unresponsive under high load on start)
blind acceptance leads to rlimit exhaustion and consequent fail with
"Too many open files" error. To prevent this we are not any more
accepting connections when their current amount is too close to
available descriptors rlimit.

Closes #3776