Fix test box-tap/cfg.test.lua on openSuSE
Found that test fails on its subtest checking:
  "gh-2872 bootstrap is aborted if vinyl_dir
   contains vylog files left from previous runs"

Debugging src/box/xlog.c found that all checks
were correct, but at function:
  src/box/vy_log.c:vy_log_bootstrap()
the check on of the errno on ENOENT blocked the
negative return from function:
  src/box/xlog.c:xdir_scan()

Found that errno was already set to ENOENT before
the xdir_scan() call. To fix the issue the errno
must be clean before the call to xdir_scan, because
we are interesting in it only from this function.
The same issue fixed at function:
  src/box/vy_log.c:vy_log_begin_recovery()

Closes #4594
Needed for #4562