Enable test reruns on failed fragiled tests
In test-run repository added ability to check per suite in suite.ini
configuration file 'fragile_retries' option, which sets the number
of accepted reruns of the test failed from 'fragile' list. Also all
suite.ini files which consisted of fragile lists were updated with
this option.

Added ability to check failed tests w/ fragile list to be sure that
the current fail equal to the issue mentioned in the fragile list.
Fragile list should consist of the results files checksums with
mentioned issues in the format:
  fragile = <basename of the test> ; gh-<issue> md5sum:<checksum>

Closes #5050