box: add has_vararg option for functions
The has_vararg option allows us to work with functions with a variable number of
arguments. This is required for built-in SQL functions.

Suppose this option is TRUE for a built-in SQL function. Then:
1) If param_list is empty, all arguments can be of any type.
2) If the length of param_list is not less than the number of the given
arguments, the types of the given arguments must be compatible with the
corresponding types described in param_list.
3) If the length of param_list is less than the number of given arguments, the
rest of the arguments must be compatible with the last type in param_list.